add todo

+++ models/User.go
@@ -4,5 +4,5 @@ type User struct {
 	UserId string
 	UserName string
 	Password string
-	CreatedAt uint16
+	CreatedAt uint64
 }
